% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{crop_region}
\alias{crop_region}
\title{Crop raster to region}
\usage{
crop_region(raster, region)
}
\arguments{
\item{raster}{SpatRaster}

\item{region}{Character region name}
}
\value{
Cropped raster
}
\description{
Crop raster to region
}
