% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{download_bioclim}
\alias{download_bioclim}
\title{Download WorldClim BIO data}
\usage{
download_bioclim(path = "data")
}
\arguments{
\item{path}{Directory to store data}
}
\value{
SpatRaster object
}
\description{
Download WorldClim BIO data
}
